requirements.txt
numpy
tensorflow-gpu
pillow
h5py
sklearn
scikit-image
pyyaml
-e .
-e tfnntools
matplotlib
jupyter
moviepy
requests